Update exposed and unexposed generator func
